feat: update design of the home page
